Add support for configuring resource requests and limits
This commit adds new parameters `controller.resources` and
`csi_driver.resources` which can be used to configure resource requests
and limits for each controller and CSI driver container.

We set fairly optimistic resource requests (currently container memory
consumption is between 15 and 25Mi), and no limits. With this
configuration, the pods won't be arbitrarily restarted due to memory
limits, and the scheduler has some information to work with.